+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/common/util/PatternStore.java
@@ -38,6 +38,12 @@ public class PatternStore {
 
     /**
      * Contains a Java-Regex-String that can be used to validate
+     * that a string is not a valid BPNL number.
+     */
+    public final static String NOT_BPNL_STRING = "^(?!(?:" + BPNL_STRING + ")$).*";
+
+    /**
+     * Contains a Java-Regex-String that can be used to validate
      * BPNA numbers.
      */
     public static final String BPNA_STRING = "^BPNA[0-9a-zA-Z]{12}$";
@@ -48,6 +54,11 @@ public class PatternStore {
      */
     public static final Pattern BPNA_PATTERN = Pattern.compile(BPNA_STRING);
 
+    /**
+     * Contains a Java-Regex-String that can be used to validate
+     * that a string is not a valid BPNA number.
+     */
+    public static final String NOT_BPNA_STRING = "^(?!(?:" + BPNA_STRING + ")$).*";
 
     /**
      * Contains a Java-Regex-String that can be used to validate
@@ -60,6 +71,12 @@ public class PatternStore {
      * It is constructed from the BPNS_STRING constant.
      */
     public static final Pattern BPNS_PATTERN = Pattern.compile(BPNS_STRING);
+
+    /**
+    * Contains a Java-Regex-String that can be used to validate
+    * that a string is not a valid BPNS number.
+    */
+    public static final String NOT_BPNS_STRING = "^(?!(?:" + BPNS_STRING + ")$).*";
 
     /**
      * Contains a Java-Regex-String that allows any combination of alphabetic characters,